pyproject.toml
[tool.poetry]
name = "loja-organicos-teste-final-squad07"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "loja_organicos_teste_final_squad07"}]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.2.2"
pandas = "^1.5.2"
pymongo = "^4.3.3"
pytest = "^7.2.0"
requests = "^2.28.1"
mysql-connector-python = "^8.0.31"
pytest-flask = "^1.2.0"
pyodbc = "^4.0.35"
fastapi = "^0.88.0"
httpx = "^0.23.1"
uvicorn = "^0.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
